backend/requirements.txt
# Backend Python dependencies
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
motor==3.6.0
pymongo==4.9.2
dnspython==2.6.1
python-dotenv==1.0.0
pydantic>=2.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
